fix: Accept strokeWidth as float, cast to int

+++ internal/conversion/gliffy.go
@@ -165,7 +165,7 @@ func AddElements(addChildren bool, input datastr.ExcalidrawScene, objects []data
 			shape.DashStyle = StrokeStyleConvExcGliffy(element.StrokeStyle)
 			shape.FillColor = FillColorConvExcGliffy(element.BackgroundColor)
 			shape.StrokeColor = element.StrokeColor
-			shape.StrokeWidth = element.StrokeWidth
+			shape.StrokeWidth = int64(element.StrokeWidth)
 			shape.Opacity = element.Opacity * 0.01
 
 			if element.FillStyle != "solid" {
@@ -208,7 +208,7 @@ func AddElements(addChildren bool, input datastr.ExcalidrawScene, objects []data
 
 				line.DashStyle = StrokeStyleConvExcGliffy(element.StrokeStyle)
 				line.StrokeColor = element.StrokeColor
-				line.StrokeWidth = element.StrokeWidth
+				line.StrokeWidth = int64(element.StrokeWidth)
 				line.FillColor = "none"
 				line.StartArrowRotation = "auto"
 				line.EndArrowRotation = "auto"
